Make `Layout` a private implementation detail of toml
