+++ src/cargo/ops/cargo_read_manifest.rs
@@ -4,23 +4,10 @@ use std::io;
 use std::path::{Path, PathBuf};
 
 use core::{Package, SourceId, PackageId, EitherManifest};
-use util::{self, paths, Config};
+use util::{self, Config};
 use util::errors::{CargoResult, CargoResultExt};
 use util::important_paths::find_project_manifest_exact;
-use util::toml::Layout;
-
-pub fn read_manifest(path: &Path, source_id: &SourceId, config: &Config)
-                     -> CargoResult<(EitherManifest, Vec<PathBuf>)> {
-    trace!("read_package; path={}; source-id={}", path.display(), source_id);
-    let contents = paths::read(path)?;
-
-    let layout = Layout::from_project_path(path.parent().unwrap());
-    let root = layout.root.clone();
-    util::toml::to_manifest(&contents, source_id, layout, config).chain_err(|| {
-        format!("failed to parse manifest at `{}`",
-                root.join("Cargo.toml").display())
-    })
-}
+use util::toml::read_manifest;
 
 pub fn read_package(path: &Path, source_id: &SourceId, config: &Config)
                     -> CargoResult<(Package, Vec<PathBuf>)> {
